+++ ic-cdk/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
 - Support VetKD management canister API. (#597)
 
 ## [0.18.0] - 2025-04-22